+++ Lab3/angelicaortizParte1-504550452.js
@@ -1,0 +1,22 @@
+
+//SIN PREGUNTAR POR UN CÓDIGO
+const libros = [
+  { id: 105, titulo: "El principito" },
+  { id: 210, titulo: "Cien años de soledad" },
+  { id: 304, titulo: "Don Quijote de la Mancha" },
+  { id: 457, titulo: "1984" },
+  { id: 501, titulo: "Un mundo feliz" },
+  // ...
+];
+
+console.log(SearchTitulo(0));
+
+function SearchTitulo(searchTitulo) {
+  for (let i = 0; i < libros.length; i++) {
+    if (libros[i].id == searchTitulo) {
+      return libros[i].titulo;
+    }
+  }
+  return "No se encuentra el titulo del libro indicado.";
+}
+
